requirements.txt
aes==1.0.0
altgraph==0.17
amqp==5.0.6
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
billiard==3.6.4.0
bootstrap4==0.1.0
boto3==1.17.93
botocore==1.20.93
bs4==0.0.1
celery==5.0.5
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
checksum==1.0.4
cipher==0.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
comtypes==1.1.7
crypto==1.4.1
cssselect==1.1.0
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
distro==1.5.0
Django==3.0.8
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-dbbackup==3.3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-maskurl==0.1
django-material==1.7.3
django-model-utils==4.1.1
django-storages==1.11.1
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.31.0
fake-useragent==0.1.11
filelock==3.0.12
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
isort==4.3.21
jmespath==0.10.0
joblib==1.0.1
jsonfield==3.1.0
kiwisolver==1.2.0
kombu==5.1.0
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupPy==1.14
matplotlib==3.3.2
mccabe==0.6.1
misaka==2.1.1
mysql==0.0.2
mysqlclient==2.0.2
Naked==0.1.31
numpy==1.19.1
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.7
panda==0.3.1
pandas==1.1.0
parse==1.18.0
pefile==2019.4.18
Pillow==7.2.0
pipenv==2020.6.2
plyer==1.4.3
prompt-toolkit==3.0.18
PyAudio @ file:///C:/Users/vatsal70/Desktop/pythonProject/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycparser==2.20
pycryptodome==3.9.8
pyee==7.0.4
pygame==2.0.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pyppeteer==0.2.2
pyquery==1.4.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-slugify==4.0.1
pyttsx3==2.90
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.66
requests==2.24.0
requests-html==0.10.0
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.6.3
seaborn==0.8.1
selenium==3.141.0
shellescape==3.8.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
swapper==1.1.2.post1
tablib==3.0.0
tabula-py==2.2.0
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.54.1
turtles==1.0.0
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.29
virtualenv-clone==0.5.4
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
whitenoise==5.2.0
wikipedia==1.4.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
